+++ speeduino/maths.h
@@ -183,9 +183,11 @@ static inline uint32_t div360(uint32_t n) {
  * @param value The value to operate on
  * @return uint32_t 
  */
-static inline uint16_t percentage(uint8_t percent, uint16_t value) {
-    return (uint16_t)div100((uint32_t)value * (uint32_t)percent);
+static inline uint32_t percentage(uint8_t percent, uint32_t value) 
+{
+    return (uint32_t)div100((uint32_t)value * (uint32_t)percent);
 }
+
 
 /**
  * @brief Integer based half-percentage calculation.
